separate table and modal
