+++ client/packages/system/src/Manage/Preferences/Components/CustomTranslations/CustomTranslationsModal.tsx
@@ -1,30 +1,14 @@
 import React, { useState } from 'react';
-import { ButtonWithIcon, DialogButton, IconButton } from '@common/components';
-import { DeleteIcon, EditIcon } from '@common/icons';
+import { ButtonWithIcon, DialogButton } from '@common/components';
+import { EditIcon } from '@common/icons';
 import {
   CUSTOM_TRANSLATIONS_NAMESPACE,
   useIntl,
   useTranslation,
 } from '@common/intl';
 import { useDialog, useToggle } from '@common/hooks';
-import {
-  Box,
-  createTableStore,
-  DataTable,
-  TableProvider,
-  TextInputCell,
-  TooltipTextCell,
-  useColumns,
-} from '@openmsupply-client/common';
-import {
-  mapTranslationsToArray,
-  mapTranslationsToObject,
-  Translation,
-} from './helpers';
-import {
-  TranslationOption,
-  TranslationSearchInput,
-} from './TranslationSearchInput';
+import { mapTranslationsToArray, mapTranslationsToObject } from './helpers';
+import { TranslationsTable } from './TranslationsInputTable';
 
 export const CustomTranslationsModal = ({
   value,
@@ -74,92 +58,6 @@ export const CustomTranslationsModal = ({
           setTranslations={setTranslations}
         />
       </Modal>
-    </>
-  );
-};
-
-const TranslationsTable = ({
-  translations,
-  setTranslations,
-}: {
-  translations: Translation[];
-  setTranslations: React.Dispatch<React.SetStateAction<Translation[]>>;
-}) => {
-  const t = useTranslation();
-
-  const columns = useColumns<Translation>([
-    {
-      key: 'key',
-      Cell: TooltipTextCell,
-      label: 'label.key',
-      width: 200,
-    },
-    {
-      key: 'default',
-      Cell: TooltipTextCell,
-      label: 'label.default',
-      width: 300,
-    },
-    {
-      key: 'custom',
-      Cell: TextInputCell,
-      label: 'label.custom',
-      cellProps: {
-        fullWidth: true,
-      },
-      setter: input => {
-        setTranslations(translations =>
-          translations.map(tr =>
-            tr.id === input.id ? { ...tr, ...input } : tr
-          )
-        );
-      },
-    },
-    {
-      key: 'delete',
-      width: 50,
-      Cell: ({ rowData }) => (
-        <IconButton
-          icon={<DeleteIcon sx={{ height: '0.9em' }} />}
-          label={t('label.delete')}
-          onClick={() =>
-            setTranslations(translations =>
-              translations.filter(tr => tr.id !== rowData.id)
-            )
-          }
-        />
-      ),
-    },
-  ]);
-
-  const onAdd = (option: TranslationOption | null) => {
-    if (!option) return;
-    const newLine = {
-      id: option.key,
-      key: option.key,
-      default: option.default,
-      custom: option.default,
-    };
-    setTranslations(translations => [...translations, newLine]);
-  };
-
-  return (
-    <>
-      <TableProvider createStore={createTableStore}>
-        <DataTable
-          id={'translations-list'}
-          columns={columns}
-          data={translations}
-          noDataMessage={t('message.add-a-translation')}
-          dense
-        />
-      </TableProvider>
-      <Box display="flex" justifyContent="flex-start" marginBottom="8px">
-        <TranslationSearchInput
-          onChange={onAdd}
-          existingKeys={translations.map(t => t.key)}
-        />
-      </Box>
     </>
   );
 };
